testing server side rendering with next.js
