+++ package.json
@@ -4,19 +4,17 @@
 	"version": "0.0.0",
 	"type": "module",
 	"scripts": {
-		"dev": "vite",
-		"build": "vite build",
+		"dev": "next dev",
+		"build": "next build",
+		"start": "next start",
 		"lint": "eslint .",
-		"format": "biome check --write",
-		"preview": "vite preview"
+		"format": "biome check --write"
 	},
 	"dependencies": {
-		"@tailwindcss/vite": "4.1.10",
+		"next": "^14.0.0",
 		"react": "^18.2.0",
 		"react-dom": "^18.2.0",
-		"react-helmet-async": "2.0.5",
 		"react-icons": "5.5.0",
-		"react-router-dom": "^7.6.2",
 		"tailwind": "4.0.0",
 		"tailwindcss": "4.1.10",
 		"typescript": "5.8.3",
@@ -27,12 +25,10 @@
 		"@eslint/js": "9.29.0",
 		"@types/react": "19.1.8",
 		"@types/react-dom": "19.1.6",
-		"@vitejs/plugin-react": "4.5.2",
 		"eslint": "9.29.0",
 		"eslint-plugin-react-hooks": "5.2.0",
 		"eslint-plugin-react-refresh": "0.4.20",
 		"globals": "16.2.0",
-		"vite": "6.3.5",
 		"wrangler": "4.20.0"
 	}
 }
