+++ src/main.tsx
@@ -1,7 +1,7 @@
 import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
 import { HelmetProvider } from "react-helmet-async";
-import App from "./App.js";
+import App from "../pages/index.js";
 
 createRoot(document.getElementById("root")!).render(
 	<StrictMode>
